consolidated state
